fix: address review feedback for PR #1064
- rename helper/test to kebab-case (server-port-editor.*)
- reuse getProxyPort from helpers.ts instead of duplicating the pref read
- extract shared waitForServerHealth into server-health.ts (used by header + editor)
- include port in the proxy-port-changed analytics event
- clarify the restart-timeout message (port is already saved)
- wire aria-describedby to the inline error; guard health-poll against unmount
